requirements.txt
django==2.2.*
pillow==6.0.*
django-debug-toolbar


## CI/CD
mypy
